fix: address CodeRabbit review comments (#2614 iter-1)
Five of six findings applied; the sixth is a false positive.

Split the loader in two (CodeRabbit). Importing `applyEnvFile` also ran
`loadEnv.ts`'s module-scope call, so `test_loadEnv.ts` read the real
repository `.env` and mutated `process.env` just by importing the helper
it tests. I had checked the blast radius — node:test isolates each file
in its own process, and the repo `.env` sets no `MULMOCLAUDE_*` key — so
nothing leaked, but "contained by luck" is not the same as "cannot
happen". The pure half now lives in `server/system/envFile.ts` with no
load-time side effect; `loadEnv.ts` is the side-effect entrypoint and
nothing but `server/index.ts` imports it.

Fix the shell check in error-recovery.md (CodeRabbit). This one was a
contradiction inside my own text: the section explains that
`export FOO=` shadows just as hard, then told the agent to run
`echo "$FOO"` — which prints blank for both unset and set-but-empty and
therefore cannot see the case being described. Now tests whether the
variable is SET (`${VAR+x}`), verified against all three states.

Split `announceShadowedEnv` (CodeRabbit). It had grown to 31 lines,
over the 20-line rule. Publishing moved to its own function, and the
stale-clear now runs on one path instead of two — with nothing shadowed,
`""` matches no id so every entry reads as stale, which is the retraction
this already wanted.

State the `DOTENV_CONFIG_*` rejection instead of only noting it is unused
(CodeRabbit). `dotenv/config` honoured PATH / OVERRIDE / ENCODING / DEBUG
/ QUIET; nothing in the repo ever set one and they were never documented,
so they are dropped deliberately rather than by omission, and the comment
now says so at the load site.

Drop the brittle line number from the e2e-live comment (CodeRabbit) — it
already pointed at the wrong line after the comment block grew.

NOT applied: `writeFileAtomic` for the test fixture write. That rule
governs server write paths, and 56 test files use `writeFileSync` against
7 that use the atomic helper — following it here would make this fixture
the outlier, not the norm.
