+++ packages/core/assets/helps/error-recovery.md
@@ -651,17 +651,20 @@ Two files can be shadowed this way — the directory the user launched from
 
 ### Fix
 
-Have the user check the shell, not the file:
+Have the user check the shell, not the file. Test whether the variable
+is **set**, not whether it prints something — `export GEMINI_API_KEY=`
+prints nothing and still shadows, which is the case `echo` cannot see:
 
 ```bash
-echo "$GEMINI_API_KEY"        # empty output ⇒ nothing is shadowing
+[ -n "${GEMINI_API_KEY+x}" ] && echo "set in the shell — this is what the app uses" \
+                             || echo "not set — the shell is not the problem"
 ```
 
-If it prints anything, that value is what the app is using. Either update
-the export to the correct value, or remove it (from the shell AND from
-`~/.zshrc` / `~/.bashrc`, or the next terminal brings it back) so the
-`.env` value takes effect. Restart the app afterwards — the load happens
-once at boot.
+If it reports "set", that shell value is what the app is using, whatever
+`.env` says. Either correct the export, or remove it — from the current
+shell AND from `~/.zshrc` / `~/.bashrc`, or the next terminal brings it
+straight back — so the `.env` value takes effect. Restart the app
+afterwards; the load happens once at boot.
 
 Do NOT tell the user to re-check the spelling in `.env`, add the key
 again, or move it elsewhere; the file is already correct, and it is being
